Refactor: useSecurityCode customHook 분리
